Add response utils
